+++ docs/change_log.rst
@@ -54,6 +54,7 @@ conversion (:trac:`14075`).
   * :func:`~stem.process.launch_tor_with_config` avoids writing a temporary torrc to disk if able (:trac:`13865`)
   * :class:`~stem.response.events.CircuitEvent` support for the new SOCKS_USERNAME and SOCKS_PASSWORD arguments (:trac:`14555`, :spec:`2975974`)
   * The 'strict' argument of :func:`~stem.exit_policy.ExitPolicy.can_exit_to` didn't behave as documented (:trac:`14314`)
+  * Threads spawned for status change listeners were never joined on, potentially causing noise during interpreter shutdown
 
  * **Descriptors**